test: moved file.test_names tests in the relevant file
