+++ tests/lib/checks/beman_standard/file/test_file.py
@@ -1,10 +1,12 @@
 #!/usr/bin/env python3
 # SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
+import pytest
 import shutil
 from pathlib import Path
 
-from beman_tidy.lib.checks.beman_standard.file import FileCopyrightCheck, FileLicenseIdCheck
+from beman_tidy.lib.checks.beman_standard.file import FileCopyrightCheck, FileLicenseIdCheck, FileTestNamesCheck
+
 
 # Workaround to test for both normal and block comments.
 test_data_prefix = Path("tests/lib/checks/beman_standard/file/data")
@@ -143,3 +145,25 @@ def test__file_license_id__fix_inplace(repo_info, beman_standard_check_config, t
 
     assert check.check() is False
     assert check.fix() is False
+
+
+# --- file.test_names tests ---
+
+test_names_prefix = Path("tests/lib/checks/beman_standard/file/data/test_names")
+
+
+def test__file_test_names__valid(repo_info, beman_standard_check_config):
+    repo_info["top_level"] = test_names_prefix / "valid"
+    check = FileTestNamesCheck(repo_info, beman_standard_check_config)
+    assert check.check() is True
+
+
+def test__file_test_names__invalid(repo_info, beman_standard_check_config):
+    repo_info["top_level"] = test_names_prefix / "invalid"
+    check = FileTestNamesCheck(repo_info, beman_standard_check_config)
+    assert check.check() is False
+
+
+@pytest.mark.skip(reason="not implemented")
+def test__file_test_names__fix_inplace(repo_info, beman_standard_check_config):
+    pass
